update cac chuc nang

+++ student-manager/database/migrations/2025_11_29_143238_create_registrations_table.php
@@ -6,24 +6,26 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-       Schema::create('registrations', function (Blueprint $table) {
-        $table->id();
-        $table->foreignId('student_id')->constrained('users')->onDelete('cascade');
-        $table->foreignId('classroom_id')->constrained('classrooms')->onDelete('cascade');
-        $table->float('score')->nullable(); // Điểm số
-        $table->unique(['student_id', 'classroom_id']); // Chống trùng
-        $table->timestamps();
-    });
+        Schema::create('registrations', function (Blueprint $table) {
+            $table->id();
+            
+            // Foreign Keys
+            $table->foreignId('student_id')->constrained('users')->onDelete('cascade');
+            $table->foreignId('classroom_id')->constrained('classrooms')->onDelete('cascade');
+            
+            // [MAX PING] Điểm số: Dùng decimal thay vì float để tránh lỗi làm tròn
+            // 4 chữ số tổng, 2 số sau dấu phẩy (VD: 10.00)
+            $table->decimal('score', 4, 2)->nullable(); 
+            
+            // Chống trùng lặp
+            $table->unique(['student_id', 'classroom_id']); 
+            
+            $table->timestamps();
+        });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('registrations');
